+++ NewsToDay/Screens/DetailedScreen/Message.swift
